Asset size_bytes is a BIGINT
The size now comes from Files.size, a Long, but the column and the row type were still 32-bit, so a video over 2 GB failed to insert.

+++ altitude/resources/migrations/postgres/all.sql
@@ -68,7 +68,7 @@ CREATE TABLE asset (
   extracted_metadata jsonb,
   folder_id CHAR(36),
   filename TEXT NOT NULL,
-  size_bytes INT NOT NULL,
+  size_bytes BIGINT NOT NULL,
   is_triaged BOOLEAN NOT NULL DEFAULT FALSE,
   is_recycled BOOLEAN NOT NULL DEFAULT FALSE,
   is_purged BOOLEAN NOT NULL DEFAULT FALSE,
