+++ altitude/src/altitude/core/dao/sql/tables/AssetRow.scala
@@ -29,7 +29,7 @@ case class AssetRow[T[_]](
     extractedMetadata: T[Option[String]],
     folderId: T[String],
     filename: T[String],
-    sizeBytes: T[Int],
+    sizeBytes: T[Long],
     isTriaged: T[Boolean],
     isRecycled: T[Boolean],
     isPurged: T[Boolean],
